Changed price type do BigDecimal in BidsEntity

+++ server/auctionapp/src/main/java/com/atlantbh/auctionapp/model/BidsEntity.java
@@ -8,6 +8,7 @@ import lombok.Setter;
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Positive;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @NoArgsConstructor
@@ -23,7 +24,7 @@ public class BidsEntity {
     @NotBlank
     @Positive(message = "value must be positive")
     @Column(nullable = false)
-    private Float price;
+    private BigDecimal price;
 
     @Column(nullable = false)
     private LocalDateTime bidDate;
@@ -36,7 +37,7 @@ public class BidsEntity {
     @JoinColumn(name = "productId", nullable = false)
     private ProductEntity product;
 
-    public BidsEntity(Float price, LocalDateTime bidDate, UserEntity user, ProductEntity product) {
+    public BidsEntity(BigDecimal price, LocalDateTime bidDate, UserEntity user, ProductEntity product) {
         this.price = price;
         this.bidDate = bidDate;
         this.user = user;
